Implement Pipeline.Empty
This is a convenience to allow creating an empty tree as part of a
pipeline. Since creating an empty tree, like any other tree, requires
repository access and therefore I/O, it could fail. This makes it a good
candidate for incorporation into a pipeline.

There is an upcoming patch to use this as part of Tree.Mkdir.

Signed-off-by: Solomon Hykes <solomon@docker.com>